Relocate phase space plot functionality away from sim
