+++ src/python/impactx/dashboard/Analyze/__init__.py
@@ -1,7 +1,9 @@
 from .analyzeFunctions import AnalyzeFunctions
 from .plot_ParameterEvolutionOverS.overS import line_plot_1d
+from .plot_PhaseSpaceProjections.phaseSpaceSettings import adjusted_settings_plot
 
 __all__ = [
+    "adjusted_settings_plot",
     "AnalyzeFunctions",
     "line_plot_1d",
 ]
